requirements.txt
flask==3.0.3
flask_login==0.6.3
PyMySQL
cryptography
flask-sqlalchemy==3.0.2
python_dotenv
apscheduler
flask_apscheduler
tzlocal==2.1
flask_cors
werkzeug==3.0.0
flask_jwt_extended==4.6.0
flask-timeloop
flask-socketio
timeloop==1.0.2
requests
redis
flask-caching
